requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
numpy==2.2.2
pdf2image==1.17.0
PyPDF2==3.0.1
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pydantic==2.10.5
python-dotenv==1.0.1
jinja2==3.1.5
aiofiles==24.1.0
httpx==0.28.1
structlog==24.4.0
google-genai>=1.0.0
